Add a new 'standalone' importer
Pursuing a new model of work here: separating out the data
about the image from the interaction with the TV.

We'll keep a separate table of all available images on disk.
Instead of relying on the TV's slideshow, we'll manage the
slideshow on the server side.

This way it becomes a lot easier to manage larger art/photo
collections by just keeping them on disk, instead of having
to do all the synchronization management of uploading everything
to the TV.

For the slideshow we'll simply upload a file and display it, and
then upload the next one, display that, and delete the previous
file.